requirements.txt
appdirs==1.4.0
cycler==0.10.0
decorator==4.0.11
imageio==2.1.2
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
